Equipment fragment should now live update and allow a User to assign their skill points to a skill of their choice.
The equipment portion of the fragment also updates in real time and will change items based on new items equipped.

EquipmentAdapter now uses a ItemDataHolder, similar to the Bestiary to store an Item Type and the Item in question assigned in that slot. When a new Item is equipped the Item property is switched with the newer Item.